working on other models

+++ packetbaseml/src/models/generic.py
@@ -33,7 +33,7 @@ class GenericModel(nn.Module):
                 inputs = inputs.to(self.device)
 
                 outputs = self(inputs)
-                loss = loss_fn(outputs, labels.to(torch.float))
+                loss = loss_fn(outputs, labels.to(torch.long))
                 loss.backward(retain_graph=True)
                 self.optim.step()
                 self.optim.zero_grad()
